Refactor ReputationManager to allow for multiple implementations
